mutex도 동적할당

+++ GoToXY.cpp
@@ -4,8 +4,8 @@
 
 void GoToXY(const int x, const int y)
 {
-	static const short originX = 20;
-	static const short originY = 25;
+	static const int originX = 20;
+	static const int originY = 25;
 
 	SetConsoleCursorPosition(GetStdHandle(STD_OUTPUT_HANDLE), COORD{ (short)(2 * (originX + x)) ,(short)(originY - y) });
 }
